pyproject.toml
[tool.poetry]
name = "Telegram Pipe"
version = "0.1.0"
description = ""
authors = ["Barahlush <baraltiva@gmail.com>"]
packages = [
    { include = "telegram_pipe"}
]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.20.2"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
pyrogram = "^2.0.97"
python-dotenv = "^0.21.1"
pyyaml = "^6.0"
types-pyyaml = "^6.0.12.3"
loguru = "^0.6.0"
tgcrypto = "^1.2.5"
uvloop = "^0.17.0"

[tool.poetry.dev-dependencies]
blue = "*"
mypy = "*"
pip-audit = "*"
pytest = "*"
pytest-cov = "*"
ruff = "*"

[tool.blue]
line-length = 79

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
strict = true

[tool.ruff]
line-length = 79
select = [
    "A",
    "ARG",
    "B",
    "C",
    "C4",
    "E",
    "ERA",
    "F",
    "I",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "RET",
    "S",
    "T10",
    "T10",
    "UP",
    "W",
]
ignore = ["A003"]
target-version = "py310"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["ARG", "S"]
